Refactor Utils
