+++ src/main/java/com/academy/dtos/service/ServiceRequestDTO.java
@@ -1,6 +1,6 @@
 package com.academy.dtos.service;
 
-import com.academy.util.FieldLengths;
+import com.academy.utils.FieldLengths;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
